+++ src/plugins/blocksmith/components/blocksmith/delegate.js
@@ -3,9 +3,9 @@ import copyBlock from '../../actions/copyBlock';
 import dirtyEditor from '../../actions/dirtyEditor';
 import storeEditor from '../../actions/storeEditor';
 
-export default (dispatch) => ({
-  handleCleanEditor: (formName) => dispatch(cleanEditor(formName)),
+export default (dispatch, { formName }) => ({
+  handleCleanEditor: () => dispatch(cleanEditor(formName)),
   handleCopyBlock: (block) => dispatch(copyBlock(block)),
-  handleDirtyEditor: (formName) => dispatch(dirtyEditor(formName)),
-  handleStoreEditor: (formName, editorState) => dispatch(storeEditor(formName, editorState)),
+  handleDirtyEditor: () => dispatch(dirtyEditor(formName)),
+  handleStoreEditor: (editorState) => dispatch(storeEditor(formName, editorState)),
 });