+++ src/plugins/news/screens/article/delegate.js
@@ -1,6 +1,10 @@
+import { validateBlocks } from '@triniti/cms/plugins/blocksmith/utils';
+import { submit } from 'redux-form';
 import AbstractDelegate from '@triniti/cms/plugins/ncr/screens/node/AbstractDelegate';
-import { formNames } from '../../constants';
+import storeEditor from '@triniti/cms/plugins/blocksmith/actions/storeEditor';
+import swal from 'sweetalert2';
 import schemas from './schemas';
+import { formNames } from '../../constants';
 
 class Delegate extends AbstractDelegate {
   constructor(dependencies) {
@@ -8,6 +12,30 @@ class Delegate extends AbstractDelegate {
       schemas,
       formName: formNames.ARTICLE,
     }, dependencies);
+  }
+
+  handleSubmit(data, formDispatch, formProps) { // eslint-disable-line consistent-return
+    const { blocksmithState, dispatch } = this.component.props;
+    const { isValid, validEditorState } = validateBlocks(blocksmithState.editorState);
+
+    if (isValid) {
+      return super.handleSubmit(data, formDispatch, formProps);
+    }
+
+    swal.fire({
+      title: 'Error saving article - one or more blocks is invalid.',
+      text: 'The invalid blocks cannot be saved. Would you like to save only the valid blocks?',
+      type: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes',
+      confirmButtonClass: 'btn btn-danger',
+      cancelButtonClass: 'btn btn-secondary',
+    }).then((result) => {
+      if (result.value) {
+        dispatch(storeEditor(formProps.form, validEditorState));
+        dispatch(submit(formProps.form));
+      }
+    });
   }
 }